+++ modules/wvp/src/main/java/com/cdzeroly/wvp/gb28181/service/IDeviceService.java
@@ -167,8 +167,8 @@ public interface IDeviceService {
     Device getDeviceBySourceChannelDeviceId(String requesterId);
 
 
-    void subscribeCatalog(int id, int cycle);
+    void subscribeCatalog(Long id, int cycle);
 
-    void subscribeMobilePosition(int id, int cycle, int interval);
+    void subscribeMobilePosition(Long id, int cycle, int interval);
 
 }
