+++ modules/wvp/src/main/java/com/cdzeroly/wvp/gb28181/service/impl/RegionServiceImpl.java
@@ -137,7 +137,7 @@ public class RegionServiceImpl implements IRegionService {
     }
 
     @Override
-    public List<RegionTree> queryForTree(String query, Integer parent, Boolean hasChannel) {
+    public List<RegionTree> queryForTree(String query, Long parent, Boolean hasChannel) {
         if (query != null) {
             query = query.replaceAll("/", "//")
                     .replaceAll("%", "/%")
