+++ modules/wvp/src/main/java/com/cdzeroly/wvp/mapper/provider/ChannelProvider.java
@@ -215,9 +215,9 @@ public class ChannelProvider {
         sqlBuild.append(BASE_SQL);
         sqlBuild.append("where channel_type = 0 and id in ( ");
 
-        Collection<Integer> ids = (Collection<Integer>)params.get("ids");
+        Collection<Long> ids = (Collection<Long>)params.get("ids");
         boolean first = true;
-        for (Integer id : ids) {
+        for (Long id : ids) {
             if (!first) {
                 sqlBuild.append(",");
             }
